backend/requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# ── HTTP client & scraping ────────────────────────────────────────────────────
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# ── Headless browser (optional) ───────────────────────────────────────────────
# playwright>=1.44.0   # uncomment if PLAYWRIGHT_ENABLED=true

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler>=3.10.0

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# ── Config / env ──────────────────────────────────────────────────────────────
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# ── Dev / testing ─────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0   # also used as async test client

# TODO: pin exact versions before first production deployment (pip-compile)
